Add early science rp omd matrices (#1748)
### Briefly, what does this PR introduce?

This PR introduces two things:

1) early-science 130 GeV RP and OMD matrices, although the OMD one is
only for testing right now (won't be used for tracks in March campaign).

2) Very mediocre protection against tracks which are truly unphysical.

These tracks come from events where the number of RP hits > 4, and one
(or more) of the hits is in one detector layer. The algorithm is not yet
able to properly sort hits to pick the "best" one.

This protection works well for events where you have a truly insane
number of hits from spallation on metal components in the RP layers, but
it also kills some "good" events with, say, 5 or 6 hits where some
simple sorting could easily through away an essentially duplicate hit in
the same layer.

The old approach did this by default for the 5-6 hit cases because it
took the first hit it found in the outer two layers.

An issue has been created to re-address this on a very short timescale
(1-2 weeks).

https://github.com/eic/EICrecon/issues/1747

### What kind of change does this PR introduce?
- [x ] Bug fix (issue #__)
- [x ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, it adds functionality.

### Does this PR change default behavior?

Yes, the overall RP tracking efficiency will drop by about 5%-10%, in
the short-term, but will no longer produce unphysical tracks.

130 GeV proton beams can now also be used with RP reconstruction for EIC
early-science.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>