remove OMD related matrices - will be put into a separate algorithm (#1752)
### Briefly, what does this PR introduce?

Fixes bug where you have conflicting matrices for 130 GeV. The problem
would only have shown up if you ran deuterons. The solution will be to
ultimately have a separate algorithm for the OMD. For now, the OMD
matrices are removed entirely.

The reason for the full removal is the addition of the 130 GeV proton
matrix. the default deuteron matrices are 130 and 135 GeV, which would
both create a conflict for using the correct matrix.

This is because the MC level information doesn't contain any indication
of a deuteron beam, so it's impossible to know at runtime which matrix
to use with a single algorithm.

Ultimately, we will need to use a separate algorithm for the OMD which
only uses the OMD hit collection, rather than sending the RP and OMD hit
collections to the same algorithm.

Closes: #1751

### What kind of change does this PR introduce?
- [x ] Bug fix (issue # 1751)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

Yes, it fixes a bug, and now OMD will not have a reconstructed track
(not a problem for the simulation campaign in the short term).

### Does this PR change default behavior?

Yes, OMD matrices are removed.

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>