Use Measurement2D for far detector clustering (#1770)
### Briefly, what does this PR introduce?
Meant as a porential alternative to
https://github.com/eic/EICrecon/pull/1765 and
https://github.com/eic/EICrecon/pull/1676 and a step towards
https://github.com/eic/EICrecon/issues/1657 without the need to change
the data model.

This allows the edm4eic::TrackerHit to remain purely a calibrated sensor
output defined by a single CellID and edm4eic::Measurement2D to be a
clustered position on a surface.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?
The type of TaggerTrackerM{}L{}ClusterPositions is changed to a
Measurement2D so will not directly contain any 3D position information.
Any downstream analysis using the collections will need to be changed.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>