LGADPulseGeneration -> SiliconPulseGeneration, LGADPulseDigitization -> EICROCDigitization (#1778)
### Briefly, what does this PR introduce?


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

LGADPulseGeneration class is removed in favor of SiliconPulseGeneration
class. Added a SiliconPulseDiscretization class to convert the smooth
pulse from SiliconPulseGeneration to it's discretized form with size of
time bin set to be equal to that of BTOF LGAD clock cycle.

LGADPulseDigitization is renamed to SiliconPulseDigitization for a
unified naming scheme.

### Does this PR change default behavior?

BTOF now uses SiliconPulseGeneration.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>