fix: remove LD_LIBRARY_PATH modifications from eicrecon-this.sh.in (#1004)
### Briefly, what does this PR introduce?
With rpath linking, much of the motivation for eicrecon-this.sh is now
obsolete and only interferes with the correct operation of other
operating system utilities.

Tests don't use eicrecon-this.sh so this is going to require some
testing by people who actually use eicrecon-this.sh in their workflows.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>