feat: separate PhotoMultiplierHitDigi readout/detector/plugin names (#1807)
### Briefly, what does this PR introduce?
The PhotoMultiplierHitDigi factory (and algorithm, to lesser extent)
assume that plugin name == detector name, and readout == plugin name +
"Hits". That's not true for the PFRICH plugin, RICHEndcapN detector,
RICHEndcapNHits readout.

This PR disentangles and corrects the detector and readout passed for
the PFRICH, avoiding an exception in the RichGeo service.

Note that this doesn't leave everything in a perfect state (the
configuration lambdas that are being set in the factory are non-ideal),
but it's the minimum amount of work to sanitize this to avoid the
exception and get the PFRICH to get digitized again (at least in theory;
I don't know if the PFRICH hits are actually stored).

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.