Add Lowq2 track association (#1775)
### Briefly, what does this PR introduce?
Reorganization of the FarDetectorLinearTracking to allow Track-Particle
associations to be produced for the Low-Q2 Tagger.

Not quite the whole picture but this PR should stand alone before the
associations are propagated to the Post-ML tracks and injected into the
common tracking flow.

There are lots of alternative approaches to this, the current one here
felt like the least ugly (assuming it works).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #1657)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Changes the output of FarDetectorLinearTracking to an edm4eic::Track
rather than edm4eic::TrackSegment. At a later date both might be
produced as the TrackSegment contains TrackPoints which are equivalent
to a missing 3D measurement.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>