fix: slight modification of maximum weight with unsigned indices (#1837)
### Briefly, what does this PR introduce?
This PR modifies the maximum weight determination in the
FarDetectorLinearTracking algorithm to allow the use of unsigned indices
and avoid compiler warnings. This has (had) a possible corner case when
all weights are zero or the number of entries is zero, when previous a
negative index would be used. These are now skipped.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: only use unsigned integers as index)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.