fix(ci): prevent uncontrolled ccache growth with 7d eviction (#1855)
### Briefly, what does this PR introduce?
It is my suspicion that the ccache is larger than needed, and it results
in evictions from the github actions cache. By doing some eviction
inside the ccache, we will hopefully limit this.

As an example, the incoming 1036 MB cache in
https://github.com/eic/EICrecon/actions/runs/14887400044/job/41810570776?pr=1855#step:4:39
is reduced to 533 MB in
https://github.com/eic/EICrecon/actions/runs/14887400044/job/41810570776?pr=1855#step:25:5.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: ccache grow too large)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.