Low-Q2 Charge Sharing (#1763)
### Briefly, what does this PR introduce?
Introduces sensor charge sharing for hits in the Low-Q2 Tagger dividing
a single SimTrackerHit into several.

The method originally introduced for LGADs has been simplified and
optimized to make it more accessible for future users, making more use
of functions provided by dd4hep.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Adds the TaggerTrackerSharedHits to the output and uses them as the
input to the digitization.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>