fix: JApplication.h -> JApplicationFwd.h (#1880)
### Briefly, what does this PR introduce?
This PR replaces includes of `JApplication.h` with `JApplicationFwd.h`
(except for the `Log_service`, where this leads to compilation
failures). The iwyu rules were updated to allow the forwarding header to
be used. JANA2 v2.2.0 includes the required forwarding header and is
already the minimum version required. Saves about 10% of compilation
time.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1879)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.