fix: move static ParticleMap into header as inline for downstream use (#1538)
### Briefly, what does this PR introduce?
Downstream plugins that use the ParticleSvc as an interface just include
the header and are supposed to link against the library to get the
static member. Having to load a shared library for just a static list is
a bit annoying for a servicde interface definition.

This PR moves the static default list into an inline static definition
(not a shared_ptr), and everything now lives in the header.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.