Make 10x100 CI run use epic_craterlake_10x100.xml (#1895)
### Briefly, what does this PR introduce?
Changes the CI run of 10x100 to use the `epic_craterlake_10x100.xml`
rather than just `epic_craterlake.xml` which is actually 5x41.

I thought historically this 10x100 was correct as the default, but it
isn't clear looking at the epic history it was ever anything else.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
CI run and report for 10x100 run will be corrected.