MatchClusters: do not assign PDG code to trackless EM clusters (#1903)
### Briefly, what does this PR introduce?
There were reports of ReconstructedParticle containing PDG=130 (K0L),
which we don't identify. Convention is that we use PID information in
Reconstructed* branches. That is already the case for tracks
(ChargedParticles), but was never addressed for when we add in clusters
in `MatchClusters`. For now the only clusters with PID are from EEEMCal,
and only electron/pion separation is implemented which requires a track
and does not belong in this branch, so we can't take that. So solution,
for now, is to keep zero values.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes, users may observe missing PDG for neutral particles reconstructed.

### Does this PR change default behavior?
Yes