Move dd4hep constant parsing from ZDC, BEMC and FHCAL plugins to algorithm init (#1892)
### Briefly, what does this PR introduce?
Moves dd4hep detector parsing out of plugins and into the algorithms.

Effected plugins are ZDC, BEMC and FHCAL.

The current approach needs to be validated by experts, particularly for
the calorimeters where the configuration naming should probably be
further changed. I'm not entirely happy with the approach for the
calorimeters, making minimal changes to try and keep the behaviour of
the current requirements the same but without generalization to a bigger
picture. Or It might be sensible to separate out the calorimeter changes
till later.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1887 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Loads detector constants though algorithms rather than plugins.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>