feat: split B0 from central detector tracking (#1744)
### Briefly, what does this PR introduce?
There is no way a relevant track can have both hits in the forward
endcaps and the B0 tracker. We can therefore separate the tracking of
both, and by doing so gain in the combinatoric search space and allow
for parallelization.

This PR splits the B0 tracker seeding and tracking from the central
detector seeding and tracking. After separate tracking, the tracks are
merged again in the `ReconstructedChargedWithoutPIDParticles`
collection.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>