Fix: Energy consistency with momentum and mass/PDG (#1928)
### Briefly, what does this PR introduce?
This resolves the issue detailed in #1919. As discussed there, for the
ReconstructedChargedParticles, this sets the energy to be consistent
with the momentum and mass/PDG.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
It will change the reconstructed energy in the
ReconstructedChargedParticles collection and the reconstructed energy
for charged particles in the ReconstructedParticles Collection.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>