SimCalorimeter Hit Contribution Time After SimCalorimeterHitProcessor Should Include Travel Time #1907 (#1923)
### Briefly, what does this PR introduce?

Currently we are (optionally) attenuation calorimeter hit contributions
in SimCalorimeterHitProcessor to be evaluated at a certain position in
the detector. However, we are not updating the time to correspond to
this attenuated signal leaving the contributions in an inconsistent
state. Moreover, this step has to be done at some point during pulse
generation. This can easily be fixed by adding this to
SimCalorimeterHitProcessor.


### This PR is currently a draft.



### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #1907)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>