updated the matrices for the OMD 130 GeV/n deuteron beams (#1957)
### Briefly, what does this PR introduce?

This PR introduces the update OMD matrices to account for changes in the
cryostat (PR: https://github.com/eic/epic/pull/893) that forced the OMD
to move further from the IP.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, this will allow PR https://github.com/eic/epic/pull/893 to be merged
(at the same time as this one) without breaking the reconstruction.

### Does this PR change default behavior?

When both PRs are merged, no, it returns things to default behavior.

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>