Add phasespace based forced pid for Far-Backward particles. (#1894)
### Briefly, what does this PR introduce?
Adds an algorithm which recovers the PID of particles detected in
certain directions. The Low-Q2 electrons are the only charged particles
which can be reconstructed within ~12mrad of the negative z-axis,
currently this information isn't included in reconstructed particles.

The same approach could to a certain degree be used for the Far-Forward
protons.

This is a temporary alternative to pursing the larger changes required
by https://github.com/eic/EICrecon/issues/1571 where the reconstructed
particles from the tagger would come from a different collection.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Tags Far-Backward reconstructed particles as electrons.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>