Added new mode to support charge sharing sigma in the units of cell dimension (#1989)
### Briefly, what does this PR introduce?

I added an option in `SiliconChargeSharing `such that `sharing_sigma
`can be defined relative to the cell size in the latest update, as per
our agreement in the reconstruction meeting. BTOF now defaults to using
this relative definition of charge sharing magnitude.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No

### Does this PR change default behavior?

No