Modified SiliconChargeSharing so it won't miss any hits (#1945)
### Briefly, what does this PR introduce?

The `SiliconChargeSharing` terminates it's iteration on
`findAllNeighborsInSensor` when `edep < m_cfg.min_edep` for a test cell.
The problem is that when the iteration ends, it will also terminate its
search for neighbors, and just because `edep < m_cfg.min_edep` for a
particular cell, that doesn't mean `edep < m_cfg.min_edep` is true for
their neighbors.

To fix this, I let `SiliconChargeSharing` to keep searching for
neighbors even for cells with `edep < m_cfg.min_edep`. `edep <
m_cfg.min_edep` is now only used to determine if a shared hits to be
created.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1944 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No

### Does this PR change default behavior?

No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>