feat: ingest EventHeader into digi, generate UID to reseed RNG (#1934)
### Briefly, what does this PR introduce?
One challenge of multi-threaded/concurrent running is that the random
number generators are called in no predictable order; this affects both
events which can come out of order, or algorithms within the same event
which can be called in different orders for different branches. That
results in non-reproducibility.

This PR ingests the event header into digitization algorithms (which is
our main use for random sampling). The event header is combined with the
algorithm name into a random seed for that event/algorithm combination.
With this seed, the random number generator is reseeded.

Outside the scope of this PR is reverting back to a single random
service instead of the various algorithm specific approaches.

Still TODO in this algorithm:
- [x] expose the top-level seed used to generate event-level seeds as a
property for setting through JANA.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: lack of reproducibility for multi-threaded
running)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes. It results in a different random number sequence, so all
distributions will be shuffled.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>