Switched to using constant fraction discriminator for BTOF sensors (#1974)
### Briefly, what does this PR introduce?

Switched to using constant fraction discriminator (`CFDROCDigitization`)
for BTOF sensors. The main branch uses constant threshold
(`EICROCDigitization`) for BTOF sensors. I was informed that EICROC was
only used on ETOF and not BTOF. As it turns out, we are using constant
fraction on BTOF after all.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No

### Does this PR change default behavior?

No

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>