feat: inherit jana::components::JOmniFactory and transition to using upstream JOmniFactory (#1959)
### Briefly, what does this PR introduce?
This PR transitions us to the upstream `jana::components::JOmniFactory`
(and generator) instead of our local version. To help the transition in
an adiabatic way, we first inherit `jana::components::JOmniFactory` into
our local JOmniFactory, and slowly transition the plugins over with
`using` switches.

This now requires JANA2 v2.3.2, which is present as of 24.10.1.
Therefore, this requires #1705.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: use upstream JOmniFactory)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>