fix: explicitly convert signed run number to unsigned for bitset (ubsan fix) (#2013)
### Briefly, what does this PR introduce?
This PR avoids an Undefined Behavior warning due to the default run
number of -1 being used in the constructor of a bitset, which only
accepts unsigned long long. Negative run numbers are static cast from -1
to 4294967295 instead of just taking abs because that would lead to -1,
0, 1 run sequences having random IDs that are identical in close
proximity.
```
SUMMARY: UndefinedBehaviorSanitizer: implicit-integer-sign-change /home/wdconinc/git/EICrecon/src/algorithms/interfaces/UniqueIDGenSvc.h:52:52 
```

See also:
-
https://github.com/JeffersonLab/JANA2/blob/a2c5bbeb00c545ed44fb7698bd9e7f1d74feb8b0/src/libraries/JANA/JFactory.h#L184
-
https://github.com/key4hep/EDM4hep/commit/5219b52264285910c65f21d7d047fb93fafd85d2

Note: no warning is suppressed for event numbers since those are not
ever -1 (and if they are, we want to know why it changed).

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>