feat(ci): create a pull request for iwyu suggestions (#1948)
### Briefly, what does this PR introduce?
This PR should make the user ergonomics for iwyu failures easier. It
creates a PR against the source branch for a PR with the changes to be
committed.
- No pipeline is started for the created PR
(https://docs.github.com/en/actions/how-tos/managing-workflow-runs-and-deployments/managing-workflow-runs/skipping-workflow-runs).
- No review is needed to merge the created PR.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: turn failing job into a concrete action)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, will create a new PR for each PR with iwyu failure.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>