Add association for TaggerTrackerReconstructedParticles, use new network with additional inputs (#1978)
### Briefly, what does this PR introduce?
Adds associations to the particles reconstructed directly from the onnx
workflow in preparation to swapping over to merging these at a later
stage of the reconstruction workflow.

Swaps to using a newer onnx model which takes in the
TaggerTrackerLocalTracks and does the projection step internally.

~~Also makes the associations and beam collection more optional.~~
Some changes have been made to make the collections optional but not
fully tested or understood if this is the approach we will make tree
wide for MC associations. A proper review will come later

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Adds missing associations which will be passed along

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>