fix: enable strongly-typed enum SiliconChargeSharingConfig::ESigmaMode (#2015)
### Briefly, what does this PR introduce?
This PR modifies the unscoped enum `ESigmaMode` to a scoped enum, and
avoids the implicit integer conversion by correcting the `operator<<`
signature. Before this fix, the output was `0` or `1` (since the
incorrectly-signatured `operator<<` is not used); after this fix, the
output is `abs` or `rel`.

The scoping of the enum adds a layer of safety since the symbol `abs`
could previously shadow the standard function `std::abs`.

This aligns with the enabled clang-tidy check
`cppcoreguidelines-use-enum-class`,
https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/use-enum-class.html.

### What kind of change does this PR introduce?
- [x] Bug fix (issue:
https://github.com/eic/EICrecon/pull/1999#discussion_r2255159290, but
fails to compile
https://github.com/eic/EICrecon/pull/1999#discussion_r2280550838)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>