fix: remove duplication of edm4eic_indexed_units in `src/algorithms/tracking` (#2062)
### Briefly, what does this PR introduce?
This PR removes a duplicated static object in `src/algorithms/tracking`,
which causes symbol clashes in unity builds, and otherwise just violates
DRY.

Note: There may be ways to provide this as part of EDM4eic as well, but
in any case we would still need to keep this around for a while here. It
would also add an Acts dependency to EDM4eic, which is hardly ideal.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>